requirements.txt
markdownify
rich
nlpia2-wikipedia
python-dotenv
replicate
requests
mdv
typer
aiohttp

